Refactor comment functionality

+++ src/main/java/com/homestore/comment/CommentRequest.java
@@ -11,5 +11,5 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class CommentRequest {
     private String text;
-    private Integer forumId;
+    private Long forumId;
 }
